+++ src/cswm.Core/Layout/Engines/BspTilingLayoutEngine.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using cswm.Core.Models;
 
 namespace cswm.Core.Layout.Engines;
