+++ src/cswm.Core/Layout/ILayoutEngine.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using cswm.Core.Models;
 
 namespace cswm.Core.Layout;
