+++ src/cswm.Core/Models/WindowInfo.cs
@@ -1,6 +1,8 @@
 using System;
 using cswm.Core.Layout;
 
+namespace cswm.Core.Models;
+
 public sealed record WindowInfo(
 	IntPtr Handle,
 	string Title,
