+++ src/cswm.Core/Services/WindowRegistry.cs
@@ -1,9 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using cswm.Core.Models;
 using Microsoft.Extensions.Logging;
 
-namespace cswm.Core.Windows;
+namespace cswm.Core.Services;
 
 public class WindowRegistry
 {
