+++ src/cswm.Core/WindowManager.cs
@@ -1,8 +1,7 @@
-using System.Collections;
 using System.Collections.Generic;
 using cswm.Core.Layout;
-using cswm.Core.Management;
-using cswm.Core.Windows;
+using cswm.Core.Models;
+using cswm.Core.Services;
 
 namespace cswm.Core;
